Started pulling ui components into their own files because apparently I hate myself.
