+++ ReQuest/cogs/admin.py
@@ -1,12 +1,14 @@
+import logging
+
 import discord
 import discord.ui
-import logging
 from discord import app_commands
 from discord.ext import commands
 from discord.ext.commands import Cog
-from ..utilities.supportFunctions import log_exception
-from ..utilities.ui import MenuDoneButton, AdminMenuButton, AdminBackButton
+
+from ..ui.views import AdminBaseView
 from ..utilities.checks import is_owner
+from ..utilities.supportFunctions import log_exception
 
 logging.basicConfig(level=logging.INFO)
 logger = logging.getLogger(__name__)
@@ -109,230 +111,6 @@ class Admin(Cog):
             await interaction.response.send_message(embed=view.embed, view=view, ephemeral=True)
         except Exception as e:
             await log_exception(e, interaction)
-
-
-class AdminBaseView(discord.ui.View):
-    def __init__(self, cdb, bot):
-        super().__init__(timeout=None)
-        self.cdb = cdb
-        self.bot = bot
-        self.embed = discord.Embed(
-            title='Administrative - Main Menu',
-            description=(
-                '__**Allowlist**__\n'
-                'Configures the server allowlist for invite restrictions.'
-            ),
-            type='rich'
-        )
-        self.add_item(AdminMenuButton(AdminAllowlistView, 'Allowlist', self.cdb, self.bot, setup_embed=False))
-        self.add_item(AdminMenuButton(AdminCogView, 'Cogs', self.cdb, self.bot,
-                                      setup_select=False, setup_embed=False))
-        self.add_item(AdminShutdownButton(bot, self))
-        self.add_item(MenuDoneButton(self))
-
-
-class AdminAllowlistView(discord.ui.View):
-    def __init__(self, cdb, bot):
-        super().__init__(timeout=None)
-        self.embed = discord.Embed(
-            title='Administration - Server Allowlist',
-            description=('__**Add New Server**__\n'
-                         'Adds a new Discord Server ID to the allowlist.\n'
-                         '**WARNING: There is no way to verify the server ID provided is valid without the bot being a'
-                         'server member. Double-check your inputs!**\n\n'
-                         '__**Remove**__\n'
-                         'Removes the selected Discord Server from the allowlist.\n\n'),
-            type='rich'
-        )
-        self.bot = bot
-        self.cdb = cdb
-        self.selected_guild = None
-        self.add_item(AdminBackButton(AdminBaseView, cdb, bot))
-
-    async def setup_select(self):
-        try:
-            self.remove_guild_allowlist_select.options.clear()
-            collection = self.cdb['serverAllowlist']
-            query = await collection.find_one()
-            if len(query['servers']) > 0:
-                for server in query['servers']:
-                    option = discord.SelectOption(label=server['name'], value=server['id'])
-                    self.remove_guild_allowlist_select.options.append(option)
-            else:
-                option = discord.SelectOption(label='There are no servers in the allowlist', value='None')
-                self.remove_guild_allowlist_select.options.append(option)
-                self.remove_guild_allowlist_select.placeholder = 'There are no servers in the allowlist'
-                self.remove_guild_allowlist_select.disabled = True
-        except Exception as e:
-            await log_exception(e)
-
-    @discord.ui.select(cls=discord.ui.Select, placeholder='Select a server to remove',
-                       options=[], custom_id='remove_guild_allowlist_select', row=0)
-    async def remove_guild_allowlist_select(self, interaction: discord.Interaction, select: discord.ui.Select):
-        try:
-            guild_id = int(select.values[0])
-            collection = self.cdb['serverAllowlist']
-            query = await collection.find_one({'servers': {'$exists': True}, 'servers.id': guild_id})
-            server = next((server for server in query['servers'] if server['id'] == guild_id))
-            logger.info(f'Found server: {server}')
-            self.selected_guild = server['id']
-            self.confirm_allowlist_remove_button.disabled = False
-            self.confirm_allowlist_remove_button.label = f'Confirm removal of {guild_id}'
-            await interaction.response.edit_message(embed=self.embed, view=self)
-        except Exception as e:
-            await log_exception(e, interaction)
-
-    @discord.ui.button(label='Add New Server', style=discord.ButtonStyle.success,
-                       custom_id='allowlist_add_server_button')
-    async def allowlist_add_server_button(self, interaction: discord.Interaction, button: discord.ui.Button):
-        try:
-            new_modal = AllowServerModal(self.cdb, self, self.bot)
-            await interaction.response.send_modal(new_modal)
-        except Exception as e:
-            await log_exception(e, interaction)
-
-    @discord.ui.button(label='Confirm', style=discord.ButtonStyle.danger, custom_id='confirm_allowlist_remove_button',
-                       disabled=True)
-    async def confirm_allowlist_remove_button(self, interaction: discord.Interaction, button: discord.ui.Button):
-        try:
-            collection = self.cdb['serverAllowlist']
-            await collection.update_one({'servers': {'$exists': True}},
-                                        {'$pull': {'servers': {'id': self.selected_guild}}})
-            await self.setup_select()
-            self.confirm_allowlist_remove_button.disabled = True
-            self.confirm_allowlist_remove_button.label = 'Confirm'
-            await interaction.response.edit_message(embed=self.embed, view=self)
-        except Exception as e:
-            await log_exception(e, interaction)
-
-
-class AllowServerModal(discord.ui.Modal):
-    def __init__(self, cdb, calling_view, bot):
-        super().__init__(
-            title='Add Server ID to Allowlist',
-            timeout=180
-        )
-        self.cdb = cdb
-        self.calling_view = calling_view
-        self.allow_server_name_input = discord.ui.TextInput(
-            label='Server Name',
-            style=discord.TextStyle.short,
-            custom_id='allow_server_name_input',
-            placeholder='Type a short name for the Discord Server',
-            required=True
-        )
-        self.allow_server_id_input = discord.ui.TextInput(
-            label='Server ID',
-            style=discord.TextStyle.short,
-            custom_id='allow_server_text_input',
-            placeholder='Type the ID of the Discord Server',
-            required=True
-        )
-        self.add_item(self.allow_server_name_input)
-        self.add_item(self.allow_server_id_input)
-        self.bot = bot
-
-    async def on_submit(self, interaction: discord.Interaction):
-        try:
-            input_name = self.allow_server_name_input.value
-            guild_id = int(self.allow_server_id_input.value)
-            collection = self.cdb['serverAllowlist']
-            self.bot.allow_list.append(guild_id)
-            await collection.update_one({'servers': {'$exists': True}},
-                                        {'$push': {'servers': {'name': input_name, 'id': guild_id}}},
-                                        upsert=True)
-            await self.calling_view.setup_select()
-            await interaction.response.edit_message(embed=self.calling_view.embed, view=self.calling_view)
-        except Exception as e:
-            await log_exception(e, interaction)
-
-
-class AdminCogView(discord.ui.View):
-    def __init__(self, cdb, bot):
-        super().__init__(timeout=None)
-        self.embed = discord.Embed(
-            title='Administration - Cogs',
-            description=(
-                '__**Load**__\n'
-                'Loads a cog by name. File must be named `<name>.py` and stored in ReQuest\\cogs\\\n\n'
-                '__**Reload**__\n'
-                'Reloads a loaded cog by name. Same naming and file path restrictions apply.'
-            ),
-            type='rich'
-        )
-        self.bot = bot
-        self.add_item(AdminBackButton(AdminBaseView, cdb, bot))
-
-    @discord.ui.button(label='Load Cog', style=discord.ButtonStyle.secondary, custom_id='admin_load_cog_button')
-    async def admin_load_cog_button(self, interaction: discord.Interaction, button: discord.ui.Button):
-        try:
-            async def modal_callback(modal_interaction: discord.Interaction, input_value):
-                module = input_value.lower()
-                await self.bot.load_extension(f'ReQuest.cogs.{module}')
-                await modal_interaction.response.send_message(f'Extension successfully loaded: `{module}`',
-                                                              ephemeral=True)
-
-            modal = AdminCogTextModal('load', modal_callback)
-            await interaction.response.send_modal(modal)
-        except Exception as e:
-            await log_exception(e, interaction)
-
-    @discord.ui.button(label='Reload Cog', style=discord.ButtonStyle.secondary, custom_id='admin_reload_cog_button')
-    async def admin_reload_cog_button(self, interaction: discord.Interaction, button: discord.ui.Button):
-        try:
-            async def modal_callback(modal_interaction: discord.Interaction, input_value):
-                module = input_value.lower()
-                await self.bot.reload_extension(f'ReQuest.cogs.{module}')
-                await modal_interaction.response.send_message(f'Extension successfully reloaded: `{module}`',
-                                                              ephemeral=True)
-
-            modal = AdminCogTextModal('reload', modal_callback)
-            await interaction.response.send_modal(modal)
-        except Exception as e:
-            await log_exception(e, interaction)
-
-
-class AdminCogTextModal(discord.ui.Modal):
-    def __init__(self, function, on_submit):
-        super().__init__(
-            title=f'{function.capitalize()} Cog',
-            timeout=180
-        )
-        self.text_input = discord.ui.TextInput(label='Name', style=discord.TextStyle.short,
-                                               placeholder=f'Enter the name of the Cog to {function}',
-                                               custom_id='cog_name_text_input', required=True)
-        self.add_item(self.text_input)
-        self._on_submit = on_submit
-
-    async def on_submit(self, interaction: discord.Interaction):
-        try:
-            await self._on_submit(interaction, self.text_input.value)
-        except Exception as e:
-            await log_exception(e, interaction)
-
-
-class AdminShutdownButton(discord.ui.Button):
-    def __init__(self, bot, calling_view):
-        super().__init__(
-            label='Shutdown',
-            style=discord.ButtonStyle.danger,
-            custom_id='shutdown_bot_button'
-        )
-        self.confirm = False
-        self.bot = bot
-        self.calling_view = calling_view
-
-    async def callback(self, interaction: discord.Interaction):
-        try:
-            if self.confirm:
-                await interaction.response.send_message('Shutting down!', ephemeral=True)
-                await self.bot.close()
-            else:
-                self.confirm = True
-                self.label = 'CONFIRM SHUTDOWN?'
-                await interaction.response.edit_message(embed=self.calling_view.embed, view=self.calling_view)
-        except Exception as e:
-            await log_exception(e)
 
 
 async def setup(bot):
